moved type and corrected imports
