+++ src/retrieve-w3c-validation/browser.ts
@@ -1,5 +1,5 @@
 // Imports
-import { W3CCSSValidatorResponse } from '..';
+import { W3CCSSValidatorResponse } from '.';
 
 // Utility function for retrieving response from W3C CSS Validator in a browser environment
 const retrieveForBrowser = async (url: string): Promise<W3CCSSValidatorResponse['cssvalidation']> => {
